Shift helpers for programming_test_cases into a separate helper
